backend/requirements.txt
fastapi
uvicorn[standard]
playwright
langchain
openai
python-dotenv
pydantic
google-generativeai
